fixing bugs, scrapping is working

+++ lordandtaylor_site/jeans_scraping/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Garment(models.Model):
     title = models.CharField(max_length=32, blank=True, default="")
     image_url = models.URLField()
-    price = models.DecimalField(decimal_places=2, max_digits=10)
+    price = models.FloatField()
     color = models.CharField(max_length=32, blank=True, default="")
     description = models.CharField(max_length=256, blank=True, default="")
